Separate out ServiceBus into its own module
